+++ app/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion 21
     buildToolsVersion '21.1.2'
     defaultConfig {
-        applicationId "com.archbrey.www.letters"
+        applicationId "com.archbrey.letters"
         minSdkVersion 16
         targetSdkVersion 22
         versionCode 1
