+++ app/src/androidTest/java/com/archbrey/letters/ApplicationTest.java
@@ -1,4 +1,4 @@
-package com.archbrey.www.letters;
+package com.archbrey.letters;
 
 import android.app.Application;
 import android.test.ApplicationTestCase;
