+++ app/src/main/AndroidManifest.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.archbrey.www.letters" >
+    package="com.archbrey.letters" >
 
     <application
         android:allowBackup="true"
-        android:icon="@mipmap/ic_launcher"
+        android:icon="@drawable/letters"
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
         <activity
