+++ app/src/main/java/com/archbrey/letters/DrawerClickListener.java
@@ -1,11 +1,10 @@
-package com.archbrey.www.letters;
+package com.archbrey.letters;
 
 import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.view.View;
-import android.widget.Adapter;
 import android.widget.AdapterView;
 import android.widget.AdapterView.OnItemClickListener;
