+++ app/src/main/java/com/archbrey/letters/DrawerLongClickListener.java
@@ -1,4 +1,4 @@
-package com.archbrey.www.letters;
+package com.archbrey.letters;
 
 import android.content.Context;
 import android.content.Intent;
