+++ app/src/main/java/com/archbrey/letters/LaunchpadActivity.java
@@ -1,4 +1,4 @@
-package com.archbrey.www.letters;
+package com.archbrey.letters;
 
 
 import android.content.BroadcastReceiver;
