+++ app/src/main/java/com/archbrey/letters/SettingsHolder.java
@@ -1,4 +1,4 @@
-package com.archbrey.www.letters;
+package com.archbrey.letters;
 
 
 /**
