+++ app/src/main/java/com/archbrey/letters/TypeOut.java
@@ -1,13 +1,11 @@
-package com.archbrey.www.letters;
+package com.archbrey.letters;
 
 import android.content.Context;
-import android.content.pm.PackageManager;
 import android.content.res.Resources;
 import android.graphics.Color;
 import android.util.TypedValue;
 import android.view.Gravity;
 import android.view.View;
-import android.widget.LinearLayout;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
